Create `CustomDimensionsWrapper`.
