+++ assets/js/modules/analytics-4/components/common/CustomDimensionsWrapper.stories.js
@@ -1,5 +1,5 @@
 /**
- * MetricTileWrapper Component Stories.
+ * CustomDimensionsWrapper Component Stories.
  *
  * Site Kit by Google, Copyright 2023 Google LLC
  *
@@ -19,22 +19,57 @@
 /**
  * Internal dependencies
  */
-import { ERROR_REASON_INSUFFICIENT_PERMISSIONS } from '../../util/errors';
-import { MODULES_ANALYTICS_4 } from '../../modules/analytics-4/datastore/constants';
-import { KM_ANALYTICS_TOP_CATEGORIES } from '../../googlesitekit/datastore/user/constants';
-import { withWidgetComponentProps } from '../../googlesitekit/widgets/util';
-import MetricTileWrapper from './MetricTileWrapper';
-import MetricTileTable from './MetricTileTable';
-import WithRegistrySetup from '../../../../tests/js/WithRegistrySetup';
-import { KEY_METRICS_WIDGETS } from './key-metrics-widgets';
+import { ERROR_REASON_INSUFFICIENT_PERMISSIONS } from '../../../../util/errors';
+import { MODULES_ANALYTICS_4 } from '../../datastore/constants';
+import { KM_ANALYTICS_TOP_CATEGORIES } from '../../../../googlesitekit/datastore/user/constants';
+import { getWidgetComponentProps } from '../../../../googlesitekit/widgets/util';
+import CustomDimensionsWrapper from './CustomDimensionsWrapper';
+import WithRegistrySetup from '../../../../../../tests/js/WithRegistrySetup';
+import { KEY_METRICS_WIDGETS } from '../../../../components/KeyMetrics/key-metrics-widgets';
+import { MetricTileTable } from '../../../../components/KeyMetrics';
 
-const WidgetWithComponentProps = withWidgetComponentProps(
-	KM_ANALYTICS_TOP_CATEGORIES
-)( MetricTileTable );
+const { Widget } = getWidgetComponentProps( KM_ANALYTICS_TOP_CATEGORIES );
 
 const Template = ( { setupRegistry, ...args } ) => (
 	<WithRegistrySetup func={ setupRegistry }>
-		<WidgetWithComponentProps { ...args } />
+		<CustomDimensionsWrapper
+			widgetSlug={ KM_ANALYTICS_TOP_CATEGORIES }
+			{ ...args }
+		>
+			<MetricTileTable
+				widgetSlug={ KM_ANALYTICS_TOP_CATEGORIES }
+				Widget={ Widget }
+				moduleSlug="analytics-4"
+				rows={ [
+					{
+						field1: [ 'keyword1' ],
+						field2: 0.112,
+					},
+					{
+						field1: [ 'keyword2' ],
+						field2: 0.212,
+					},
+					{
+						field1: [ 'keyword3' ],
+						field2: 0.312,
+					},
+				] }
+				columns={ [
+					{
+						field: 'field1.0',
+						Component: ( { fieldValue } ) => (
+							<a href="http://example.com">{ fieldValue }</a>
+						),
+					},
+					{
+						field: 'field2',
+						Component: ( { fieldValue } ) => (
+							<strong>{ fieldValue }</strong>
+						),
+					},
+				] }
+			/>
+		</CustomDimensionsWrapper>
 	</WithRegistrySetup>
 );
 
@@ -84,7 +119,7 @@ Ready.parameters = {
 export const ErrorMissingCustomDimensions = Template.bind( {} );
 ErrorMissingCustomDimensions.storyName = 'Error - Missing custom dimensions';
 ErrorMissingCustomDimensions.scenario = {
-	label: 'KeyMetrics/MetricTileWrapper/ErrorMissingCustomDimensions',
+	label: 'KeyMetrics/CustomDimensionsWrapper/ErrorMissingCustomDimensions',
 };
 ErrorMissingCustomDimensions.parameters = {
 	features: [ 'newsKeyMetrics' ],
@@ -115,7 +150,7 @@ ErrorCustomDimensionsInsufficientPermissions.args = {
 	},
 };
 ErrorCustomDimensionsInsufficientPermissions.scenario = {
-	label: 'KeyMetrics/MetricTileWrapper/ErrorCustomDimensionsInsufficientPermissions',
+	label: 'KeyMetrics/CustomDimensionsWrapper/ErrorCustomDimensionsInsufficientPermissions',
 };
 ErrorCustomDimensionsInsufficientPermissions.parameters = {
 	features: [ 'newsKeyMetrics' ],
@@ -146,17 +181,16 @@ ErrorCustomDimensionsGeneric.args = {
 	},
 };
 ErrorCustomDimensionsGeneric.scenario = {
-	label: 'KeyMetrics/MetricTileWrapper/ErrorCustomDimensionsGeneric',
+	label: 'KeyMetrics/CustomDimensionsWrapper/ErrorCustomDimensionsGeneric',
 };
 ErrorCustomDimensionsGeneric.parameters = {
 	features: [ 'newsKeyMetrics' ],
 };
 
 export default {
-	title: 'Key Metrics/MetricTileWrapper',
-	component: MetricTileWrapper,
+	title: 'Key Metrics/CustomDimensionsWrapper',
+	component: CustomDimensionsWrapper,
 	args: {
-		moduleSlug: 'analytics-4',
 		setupRegistry: () => {},
 	},
 };
