+++ assets/js/modules/analytics-4/components/common/CustomDimensionsWrapper.test.js
@@ -1,5 +1,5 @@
 /**
- * Key Metrics MetricTileWrapper component tests.
+ * Key Metrics CustomDimensionsWrapper component tests.
  *
  * Site Kit by Google, Copyright 2023 Google LLC
  *
@@ -20,22 +20,23 @@ import {
 	provideUserAuthentication,
 	provideUserCapabilities,
 	render,
-} from '../../../../tests/js/test-utils';
-import { withWidgetComponentProps } from '../../googlesitekit/widgets/util';
-import { ERROR_REASON_INSUFFICIENT_PERMISSIONS } from '../../util/errors';
-import { KEY_METRICS_WIDGETS } from './key-metrics-widgets';
-import { KM_ANALYTICS_TOP_CATEGORIES } from '../../googlesitekit/datastore/user/constants';
-import { MODULES_ANALYTICS_4 } from '../../modules/analytics-4/datastore/constants';
-import MetricTileTable from './MetricTileTable';
+} from '../../../../../../tests/js/test-utils';
+import { getWidgetComponentProps } from '../../../../googlesitekit/widgets/util';
+import { ERROR_REASON_INSUFFICIENT_PERMISSIONS } from '../../../../util/errors';
+import { KEY_METRICS_WIDGETS } from '../../../../components/KeyMetrics/key-metrics-widgets';
+import { KM_ANALYTICS_TOP_CATEGORIES } from '../../../../googlesitekit/datastore/user/constants';
+import { MODULES_ANALYTICS_4 } from '../../datastore/constants';
+import MetricTileTable from '../../../../components/KeyMetrics/MetricTileTable';
+import CustomDimensionsWrapper from './CustomDimensionsWrapper';
 
-describe( 'MetricTileWrapper', () => {
-	const WidgetWithComponentProps = withWidgetComponentProps(
-		KM_ANALYTICS_TOP_CATEGORIES
-	)( MetricTileTable );
+describe( 'CustomDimensionsWrapper', () => {
+	const { Widget } = getWidgetComponentProps( KM_ANALYTICS_TOP_CATEGORIES );
 
 	it( 'renders appropriate error if required custom dimensions are not available', () => {
 		const { container } = render(
-			<WidgetWithComponentProps moduleSlug="analytics-4" />,
+			<CustomDimensionsWrapper
+				widgetSlug={ KM_ANALYTICS_TOP_CATEGORIES }
+			/>,
 			{
 				setupRegistry: ( registry ) => {
 					provideUserAuthentication( registry );
@@ -58,7 +59,9 @@ describe( 'MetricTileWrapper', () => {
 
 	it( 'renders appropriate error if creating custom dimensions failed due to insufficient permissions', () => {
 		const { container } = render(
-			<WidgetWithComponentProps moduleSlug="analytics-4" />,
+			<CustomDimensionsWrapper
+				widgetSlug={ KM_ANALYTICS_TOP_CATEGORIES }
+			/>,
 			{
 				setupRegistry: ( registry ) => {
 					provideUserAuthentication( registry );
@@ -96,7 +99,9 @@ describe( 'MetricTileWrapper', () => {
 
 	it( 'renders appropriate error if creating custom dimensions failed due to a generic error', () => {
 		const { container } = render(
-			<WidgetWithComponentProps moduleSlug="analytics-4" />,
+			<CustomDimensionsWrapper
+				widgetSlug={ KM_ANALYTICS_TOP_CATEGORIES }
+			/>,
 			{
 				setupRegistry: ( registry ) => {
 					provideUserAuthentication( registry );
@@ -134,37 +139,41 @@ describe( 'MetricTileWrapper', () => {
 
 	it( 'renders report correctly if there are no errors', () => {
 		const { container } = render(
-			<WidgetWithComponentProps
-				moduleSlug="analytics-4"
-				rows={ [
-					{
-						field1: [ 'keyword1' ],
-						field2: 0.112,
-					},
-					{
-						field1: [ 'keyword2' ],
-						field2: 0.212,
-					},
-					{
-						field1: [ 'keyword3' ],
-						field2: 0.312,
-					},
-				] }
-				columns={ [
-					{
-						field: 'field1.0',
-						Component: ( { fieldValue } ) => (
-							<a href="http://example.com">{ fieldValue }</a>
-						),
-					},
-					{
-						field: 'field2',
-						Component: ( { fieldValue } ) => (
-							<strong>{ fieldValue }</strong>
-						),
-					},
-				] }
-			/>,
+			<CustomDimensionsWrapper widgetSlug={ KM_ANALYTICS_TOP_CATEGORIES }>
+				<MetricTileTable
+					widgetSlug={ KM_ANALYTICS_TOP_CATEGORIES }
+					Widget={ Widget }
+					moduleSlug="analytics-4"
+					rows={ [
+						{
+							field1: [ 'keyword1' ],
+							field2: 0.112,
+						},
+						{
+							field1: [ 'keyword2' ],
+							field2: 0.212,
+						},
+						{
+							field1: [ 'keyword3' ],
+							field2: 0.312,
+						},
+					] }
+					columns={ [
+						{
+							field: 'field1.0',
+							Component: ( { fieldValue } ) => (
+								<a href="http://example.com">{ fieldValue }</a>
+							),
+						},
+						{
+							field: 'field2',
+							Component: ( { fieldValue } ) => (
+								<strong>{ fieldValue }</strong>
+							),
+						},
+					] }
+				/>
+			</CustomDimensionsWrapper>,
 			{
 				setupRegistry: ( registry ) => {
 					provideUserAuthentication( registry );
