+++ assets/js/modules/analytics-4/components/common/index.js
@@ -20,3 +20,4 @@ export { default as EnhancedMeasurementSwitch } from './EnhancedMeasurementSwitc
 export { default as WebDataStreamSelect } from './WebDataStreamSelect';
 export { default as PropertySelect } from './PropertySelect';
 export { default as UseSnippetSwitch } from './UseSnippetSwitch';
+export { default as CustomDimensionsWrapper } from './CustomDimensionsWrapper';
