+++ tests/backstop/reference/google-site-kit_KeyMetrics_CustomDimensionsWrapper_ErrorCustomDimensionsGeneric_0_document_2_large.png
