+++ tests/backstop/reference/google-site-kit_KeyMetrics_CustomDimensionsWrapper_ErrorCustomDimensionsInsufficientPermissions_0_document_2_large.png
