+++ tests/backstop/reference/google-site-kit_KeyMetrics_CustomDimensionsWrapper_ErrorMissingCustomDimensions_0_document_0_small.png
