+++ tests/backstop/reference/google-site-kit_KeyMetrics_CustomDimensionsWrapper_ErrorMissingCustomDimensions_0_document_2_large.png
