+++ tests/backstop/reference/google-site-kit_KeyMetrics_CustomDimensionsWrapper_Ready_0_document_1_medium.png
